requirements.txt
# 股票数据采集系统依赖包

# 数据采集相关
akshare==1.17.52
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# 数据存储相关
elasticsearch==9.1.0

# 消息队列相关
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# 流处理相关
apache-flink==1.18.1

# 任务调度相关
schedule>=1.2.0
apscheduler>=3.10.4
Flask-APScheduler>=1.13.1

# 工具库
python-dotenv>=1.0.0
loguru>=0.7.0
pytz>=2023.3
jsonschema>=4.17.0
retrying>=1.3.4

# Web框架和API
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-RESTful>=0.3.10
Werkzeug>=2.3.0
gunicorn>=21.0.0

# API文档和验证
marshmallow>=3.20.0
flasgger>=0.9.7

# 限流和缓存
Flask-Limiter>=3.5.0
redis>=4.6.0